requirements.txt
google-api-python-client
oauth2client
google.auth
grpcio == 1.46.3
google-cloud-monitoring

